Move getUsernameSuggestion function and method from server/ to rc-lib
