+++ packages/rocketchat-lib/server/functions/index.js
@@ -23,3 +23,4 @@ export { setUsername } from './setUsername';
 export { unarchiveRoom } from './unarchiveRoom';
 export { updateMessage } from './updateMessage';
 export { validateCustomFields } from './validateCustomFields';
+export { generateUsernameSuggestion } from './getUsernameSuggestion';
