+++ packages/rocketchat-lib/server/index.js
@@ -1,0 +1,2 @@
+export * from './lib';
+export * from './functions';
